// frontend/src/lib/api/generated/model/githubComAbhishekPenDriveBackendInternalApiDtoFileSystemEntry.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * pen-drive API
 * pen-drive backend API
 * OpenAPI spec version: 0.1.0
 */

export interface GithubComAbhishekPenDriveBackendInternalApiDtoFileSystemEntry {
  last_modified?: string;
  name?: string;
  path?: string;
  size?: number;
  type?: string;
}